chore: isolate table row and column mutations
